+++ app/data/spells/water.ts
@@ -1,0 +1,106 @@
+import { Spell } from "../../types/spell.ts";
+
+export const waterSpell: Array<Spell> = [
+  {
+    id: "17",
+    name: "Cure",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "18",
+    name: "Dispel",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "19",
+    name: "Bless",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "20",
+    name: "Protection from Water",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "21",
+    name: "Summon Boat",
+    level: 1,
+    effect: "",
+  },
+  {
+    id: "22",
+    name: "Ice Bolt",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "23",
+    name: "Remove Obstacle",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "24",
+    name: "Scuttle Boat",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "25",
+    name: "Weakness",
+    level: 2,
+    effect: "",
+  },
+  {
+    id: "26",
+    name: "Forgetfulness",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "27",
+    name: "Frost Ring",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "28",
+    name: "Mirth",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "29",
+    name: "Teleport",
+    level: 3,
+    effect: "",
+  },
+  {
+    id: "30",
+    name: "Clone",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "31",
+    name: "Prayer",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "32",
+    name: "Water Walk",
+    level: 4,
+    effect: "",
+  },
+  {
+    id: "33",
+    name: "Summon Water Elemental",
+    level: 5,
+    effect: "",
+  },
+];